backend/requirements.txt
# ===== FastAPI =====
fastapi==0.115.14
uvicorn[standard]==0.35.0

# ===== Database =====
SQLAlchemy==2.0.41
psycopg[binary]==3.2.3
alembic==1.13.2

# ===== Pydantic =====
pydantic==2.11.7
pydantic-settings==2.10.1
email-validator==2.2.0

# ===== Auth =====
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4

# ===== Environment =====
python-dotenv==1.1.1

# ===== File upload =====
python-multipart==0.0.20

# ===== Utils =====
python-dateutil==2.9.0.post0